deletion requires prior confirmation

+++ standalone/src/main/java/org/jboss/as/console/client/domain/hosts/ServerConfigView.java
@@ -100,7 +100,8 @@ public class ServerConfigView extends SuspendableViewImpl implements ServerConfi
                         new Feedback.ConfirmationHandler() {
                             @Override
                             public void onConfirmation(boolean isConfirmed) {
-                                presenter.deleteCurrentRecord();
+                                if(isConfirmed)
+                                    presenter.deleteCurrentRecord();
                             }
                         });
             }
